build: compartmentalize
